+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Main
+
+- Replace "End in period" suggestion with "End in punctuation". [#49](https://github.com/double-great/alt-text/pull/49)
+
 ## 0.3.1
 
 - Register "Avoid emoji" suggestion with main script. [#48](https://github.com/double-great/alt-text/pull/48)